backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
firebase-admin==6.3.0
google-cloud-firestore==2.14.0
httpx==0.25.2
python-docx==1.1.0
python-pptx==0.6.23
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
better-profanity==0.7.0
beautifulsoup4==4.12.2
